SWARM-654: Hollow jars are not recognized, not deploying
Motivation
----------
With the switch to using ApplicationEnvironment it looks like some previous behavior wasn't updated. This results in Hollow jars being generated with an `asset` in the YAML even though there is no deployment in /bootstrap of the uber jar.

Also, deployments are not recognized on the command line as its trying to deploy a default deployment in the hollow jar, which doesn't exist

Modifications
-------------
Don't set asset and application artifact values in YAML when hollow is true, and use ApplicationEnvironment.isHollow() to determine what Swarm.deploy() should do.

Result
------
Fixes broken hollow jar creation and properly deploys the correct artifact.