SWARM-667: Dependencies not found in Maven Profile (#141)
Motivation
----------
When WildFly Swarm dependencies are present only in a Maven Profile, they are unable to be found when building the container for Arquillian tests.

Modifications
-------------
Check for presence of `env.MAVEN_CMD_LINE_ARGS` which contains the Maven command line. Then extract the parts relating to profiles to pass it into ShrinkwrapResolver when loading pom.xml.

Result
------
When dependencies are present in a Maven profile, Arquillian tests will find them when run on the command line.