SWARM-1 - Allow packaging=jar to deploy web things.
Also, SWARM-682 and SWARM-362, which are intertwinglyly related.

Motivation
----------

For simple applications, you shouldn't have to build a .war

Modifications
-------------

Now, if :undertow or :jaxrs is present, and the default deployment
happens to be a .jar, we think that is obviously an error, and we
strongly coerce that .jar into a servlet or jaxrs war by re-arranging
the classes, adding dependencies, etc.

Additionally, the dependency between :cdi and :undertow has been
broken as it's not actually required.

Result
------

Now you can produce a normal .jar project and magically get a .war
deployed.  Only worthwhile if you're doing the default deployment.

Additionally, you can have a "pure" CDI app that doesn't involve
any web components now.