SWARM-690: FractionList dependencies have incomplete information
Motivation
----------
Some metadata from FractionDescriptors are incorrect when a fraction (eg. archaius) is not present in the root
but declared only as a dependency.

Modifications
-------------
The fraction list parsing code was optimized to use the complete information provide in the JSON file only.

Result
------
The correct metadata is returned for the specified fraction