SWARM-697: Fix StackOverflowError during gradle artifactresolving
Motivation
----------
There are 3rd-party libs (even on mavencentral) which have circular dependencies. This shouldn't break the build when building the swarm-jar. It should also be possible to force versions of transitive dependencies.

Modifications
-------------
Used gradle-features to resolve artifacts to a given configuration. This is much faster and can handle the circular dependencies. The detached configuration to resolve dependencies copies the forced modules from the compile-configuration.

Result
------
The wildfly-swarm-package task can handle circular 3rd-party libs and respects forced versions from the ResolutionStrategy from the configuration compile.