SWARM-437 - Allow annotations to advertise services. (#154)
Motivation
----------

In the case a person doesn't use a main(), and wants to perform
multiple advertisements, particularly those not based on the simplified
archive name, provide an @Advertise annotation.

Modifications
-------------

@Annotation is inspected for each archive, and can specify the
name of services to advertise.  It is a repeatable annotation
and several can be applied to any class, for whatever reason.

This commit also reverses the apparently arbitrary choice we made
to run the ArchivePreparers before the MetadataProcessors.

Result
------

@Advertise("tacos")
public class Tacos { ... }

Will cause the same behaviour as calling

  archive.as(TopologyArchive.class).advertise("tacos")

on the deployed archive containing the aforementioned class.