SWARM-543 & SWARM-721 - Remote MQ Connections
Motivation
----------

While the WildFly ActiveMQ/Artemis subsystem supports
remote connections to remote servers, we did not expose
enough to provide configuration, nor document it well.

Modifications
-------------

Expose, for SWARM-721 -

  swarm.socketBinding(...)
  swarm.outboundSocketBinding(...)

Which allow for user-centric provisioning of additional
socket-bindings and outbound-socket-bindings.

Additionally, provide a test (as an example) of how to use
them, along with configuring the MessagingFraction, to connect
to an external ActiveMQ/Artemis server.

Result
------

A remote connection-factory is now available in the pooled
connection pool, and can also be used for MDBs if the
correct @ResourceAdapter is used.

Further information available:

https://docs.jboss.org/author/display/WFLY10/Connect+a+pooled-connection-factory+to+a+Remote+Artemis+Server