SWARM-438 - Programttic API for advertising services. (#158)
Motivation
----------

In the case where a user desires run-time advertisement of
services, as opposed to using archive.advertise() or the
@Advertise annotation, we should provide a way to accomplish
it.

Modifications
-------------

Added a method to the Topology interface, which can be obtained
through JNDI.  topology.advertise(name) will advertise for
http and https (if available) for the current IP address of the
process, the given service-name provided as an argument.

The return value is a handle which may be used to unadvertise()
the service.  Once unadvertised, the service may not be re-used
without obtaining a new handle.

Result
------

A user may now, from within a running deployment, advertise
arbitrary services.