SWARM-726 - Easier configuration of remote MQ connections.
Motivation
----------

While we now (as of recently) support remote MQ connections, they
were very in-depth to configure, and exposed a little too much
implementation detail of WildFly in order to simply connect to
an outboard MQ.

Modifications
-------------

Our EnhancedServer now provides a variety of methods:

  .remoteConnection() // completely default, localhost:61616, java:/jms/remote-mq
  .remoteConnection(name) // completely default, aside from name/jndi
  .remoteConnection(connection) // completely customized
  .remoteConnection(connectionSupplier) // completely customized
  .remoteConnection(name, connectionConsumer) // completely customized

Additionally, in the case you only need exactly 1 external remote
connection, it can be configured via configuration values
(via properties or YAML).

  swarm.messaging.remote (a boolean-ish value)
  swarm.messaging.remote.name
  swarm.messaging.remote.jndi-name
  swarm.messaging.remote.host
  swarm.messaging.remote.port

Where the jndi-name is usually derived from the name, but separately
overridable.  Host and port have defaults (localhost and 61616 respectively).

To enable config-value configuration of a remote connection, at least
one of the above keys must be set in order to trigger the auto-setup.

Result
------

Significantly easier to setup a remote MQ connection.