SWARM-582: saaj-impl module not found
Motivation
----------
Within the webservices subsystem of WildFly it directly loads this module when it's unable to find a service entry for javax.xml.soap.MetaFactory.

As its directly loading it, our plugin doesn't have a way to know it needs that module present so we need to specifically reference it in our build.

Modifications
-------------
Added a test showing the problem and then updated module.conf for webservices fraction to include saaj-impl

Result
------
Solves exception when trying to lookup a MetaFactory for SOAP.