SWARM-742: default initial logging format is different from default logging subsystem format
Motivation
----------
When running a -swarm.jar application with default logging
configuration, there's a visible difference in log message
format before and after the logging subsystem kicks in.

Modifications
-------------
Change the default logging.properties configuration to be
identical to the default LoggingFraction configuration.

Result
------
With default logging configuration, there will be no visible
difference in the log messages before and after the logging
subsytem kicks in.