SWARM-671 - Behave better when `mvn wildfly-swarm:package` used on commandline.
Motivation
----------

In the event a user called `mvn wildfly-swarm:package` from the
commandline, and not via a pom.xml, we should try harder to do
the right thing (produce a -swarm.jar) or provide a better error
message if we can't.

Modifications
-------------

Altered the plugin to fallback to trying to divine the input
primary artifact file even if the entire maven build-cycle has
not fired, by looking for the correct (best guess) file within
the output directory.

From a virgin build, `mvn wildfly-swarm:package` still does not
succeed, but at least it has a better chance of succeeding.

Result
------

If a user has not included the execution wildfly-swarm-plugin in
their POM, but still manage to invoke it from the commandline, one
of two things will happen:

* If their .jar or .war has been previously built, we will find
  it an build a -swarm.jar as desired.

* If their .jar or .war has _not_ been previously built, we will
  produce an error message akin to:

    Cannot package without a primary artifact;
    please `mvn package` prior to invoking
    wildfly-swarm:package from the command-line