SWARM-756: Restructure project modules
Motivation
----------
It has become increasingly difficult to know what the different modules of the project provide, so its necessary to add some delineation between areas

Modifications
-------------
We now have /core for the main runtime setup pieces.
/fractions contains all the fractions, which is further subdivided into
/fractions/javaee
/fractions/netflix
/fractions/camel
/fractions/cdi-extensions

As part of this cdi-ext has moved to /fractions/cdi-extensions/cdi-config and been renamed to cdi-config.

Result
------
No impact on runtime behavior, simply a re-org of pieces for development