SWARM-775 - Convert our Arquillian support to a fraction.
Motivation
----------

Our Arquillian support has been an odd-duck, particularly
when trying to get and keep the daemon running.

Also, @InVM is a morass of crap.

Modifications
-------------

Convert to a fraction where the fraction itself starts
and manages the daemon, and the only thing we place into
the archive appender is the bit that connects the deployment
to the daemon.

Also, Remove @InVM.

Result
------

Nothing visibly different, except @InVM is gone.