SWARM-778 - Simplify @Deployment creation for ARQ tests
Motivation
----------

Creating the @Deployment for at least our internal ARQ tests
is a pain. It involves a lot of explicit addition of classes
and resources.  It could be simpler.

Modifications
-------------

Provide a class-level @DefaultDeployment annotation which can
be used (optionally) instead of @Deployment.  Takes the same
testable=(true|false) argument, but if found, tries its best
to JFDI in constructing the deployment.  This involves basically
finding anything near the test-class, and jamming it into a .jar.

Currently .war isn't explicitly supported.

As proof-of-concept, the Batch-JBeret and CDI testsuites
have been converted to use @DefaultDeployment, along with a
fair bit of code deletion.

Result
------

Simpler and more straight-forward of ARQ testing, at least for
our own internal bits.