SWARM-781 - Implicitly add beans.xml if required.
Motivation
----------

A user includes our :cdi fraction, they shouldn't also be
required to ensure there's an empty beans.xml.

But if they do provide one, we should respect it.

Modifications
-------------

An archive-preparer that checks for and adds a beans.xml
if one is missing from the deployment, if the CDI fraction
is involved.

Result
------

No more requirement of an empty beans.xml.