SWARM-809 - Better waiting for Swarm processes to start.
Motivation
----------

In some examples our tests would be fooled by the first
deployed item, which sometimes was an implicit or otherwise
non-primary deployment.  This caused tests to run before
the actual code-to-be-tested was ready.

Modifications
-------------

Added a hook to display a better log message once the
default or user-provided main() had completed its work
and wire that to SwarmProcess.awaitReadiness() instead
of watching for any particular deployment.

Result
------

Better testing.