SWARM-739 (and others) Improve configuration.
Motivation
----------

Configuration is a mish-mash of values from fractions, properties
and YAML files.  Defaults are scattered to the winds. This is sub-optimal.

Modifications
-------------

Replace many fraction instance attributes with newly-created
Configurable<T>, which respect the idea of YAML wins over
user-set-on-Fraction wins over defaults.

Lessens the need to do specific value inspection and logic
within Customizers and such.

Configurable<T> can currently be used (and discovered) in:

  - Fractions
  - Customizers
  - ArchivePreparers
  - ArchiveMetadataProcessors

Usually they will live within a Fraction but that is not always
appropriately.

Result
------

Cleaner and more logical configuration, simpler customizers. More
information logged when a process starts.