SWARM-739 (and others) Improve configuration. (#222)
Motivation
----------

Configuration is a mish-mash of values from fractions, properties
and YAML files.  Defaults are scattered to the winds. This is sub-optimal.

Modifications
-------------

Introduce a Defaultable<T> for providing fancier and trackable
defaults. Not used terribly many places, but in a few Customizers
and enhanced versions of fraction resources.  Currently used
for indicating that one default value is dependent upon some
other value, such as jndi-name=java:/jms/%{name} or such.

Additionally, provide a @Configurable annotation, to denote
that some arbitrary field should be configurable through the
project-stages/system-properties method, and to specify the
name to bind to that field.

@Configurable will be detected upon the following classes

  - Fractions
  - Customizers
  - ArchivePreparers
  - ArchiveMetadataProcessors

Additionally, any Fraction implementation is deemed to be
*implicitly* configurable, and the root Fraction object is
walked through its entire subresources to discover and name
configurable items.

In addition to allowing overriding values set by default
or through a main() using the project-stages.yml file,
completely new items may be created simply through the
specification of the appropriate configuration values.

e.g.

  swarm:
    datasources:
      data-sources:
        MyDS:
          driver-name: myh2
          connection-url: jdbc:h2:mem:test;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
          user-name: sa
          password: sa
      jdbc-drivers:
        myh2:
          driver-class-name: org.h2.Driver
          xa-datasource-name: org.h2.jdbcx.JdbcDataSource
          driver-module-name: com.h2database.h2

Is exactly the equivalent of:

  datasourcesFraction.dataSource("MyDS");
  datasourcesFraction.jdbcDriver("myh2");

followed by the abovementioned setting of properties within
the fraction tree as though the user had invoked the above
two lines before applying configuration values.  In the
event that .jdbcDriver(key, consumer) exists and applies
extra logic (in the case of drivers, for instance, we set
a default driver-name to be equivalent to the key for the
resource via DatasourceFraction), that logic still applies,
hopefully resulting in unsurprising behavior.

Result
------

Cleaner, more logical and more consistent configuration, simpler customizers.
More information logged when a process starts.