SWARM-846: TopologyProxyService is never started
Motivation
----------
TopologyWebAppActivator is not triggering the installation of the TopologyProxyService which means there is no way to map between a proxy URL and the underlying topology service when a main() uses TopologyWebAppFraction.proxyService().

This bug currently prevents Booker from looking up services.

Modifications
-------------
Modified TopologyWebAppActivator to be picked up by Weld and added as a ServiceActivator into runtime.

TopologyProzyService was creating service name with .of() instead of .parse() which caused quotes to be part of service name.

Modified topology-webapp test to check for presence of appropriate MSC services related to proxies.

Result
------
Product behavior corrected from failing silently.