SWARM-823 - Bring-your-own JDBC driver (#224)
Motivation
----------

With our fractions bringing in (via transitivity) particularly
versions of JDBC drivers, and wrapping them in a module.xml,
it's quite difficult to keep up to date, and allow users to
bring alternate versions of these drivers.

Modifications
-------------

Teach WildFly Swarm to not transitively bring in any drivers,
but able to recognize them.  Upon recognition, it can create
a synthetic module.xml and register the driver, regardless
of the actual version of the driver jar is involved.

Result
------

Anything that previously assumed h2 was present will need to
explicitly add a dependency on the h2 JDBC driver jar (with
scope=compile).  Any version is fine.

Auto-detectable (in theory) databases include:

* H2
* MySQL
* PostgreSQL
* DB2
* EnterpriseDB
* SQLServer
* Sybase
* Oracle