SWARM-857: Secured Ribbon client CNFE
Motivation
----------
Using the secured Ribbon client gets a ClassNotFoundException for KeycloakSecurityContextAssociation when attempting to propagate the Bearer Token from Keycloak to a subsequent call.

This results in the security context not being present on a subsequent call.

Modifications
-------------
ribbon.secured:client module incorrectly had a dependency on swarm.keycloak:runtime instead of :deployment.

Result
------
No longer an exception when propagating bearer token between calls and security context is available.