SWARM-810: Update OpenShift Client
Motivation
----------
Update the version of the OpenShift Client being used and modify the topology to add a watch as opposed to spawning a Thread that polls.

Modifications
-------------
Updated module.xml dependencies for OpenShift Client library and modified ServiceWatcher to add itself as a listener to events from OpenShift.

Updates the topology based on the events for Kubernetes Services Add/Change, and uses the service name when registering it into topology instead of the IP.

Result
------
No impact to product behavior