SWARM-862: Multi Module Api from testsuite not in M2 repo
Motivation
----------
When executing `mvn dependency:tree` it's unable to find multi-module-api artifact when trying to build multi-module-view.

This is because multi-module-api is not installed into the M2 repo during a build.

Modifications
-------------
Cleaned up some formatting and overrode the skipping of the installation for multi-module-api artifact.

Result
------
No impact on product, purely testsuite related.