SWARM-858: additional fractions: unintuitive behavior if the fraction is not found (#235)
Motivation
----------
If the user sets an additional fraction in the Maven plugin
configuration but makes a mistake, two things can happen:

1. If the user used the format "artifactId", an unhelpful
   error message is shown.
2. If the user used the format "artifactId:version"
   or "groupId:artifactId:version", build passes just fine.

Both of these behaviors are unintuitive.

Modifications
-------------
The method FractionDescriptor.fromGav was changed to always
consult the FractionList. If the fraction isn't found
in the list, an appropriate error message is used.

Result
------
Wrong configuration of additional fractions will now always
result in build errors.