SWARM-903: Display a warning when HTTP/2 is not supported in an environment
Motivation
----------
HTTP/2 may not work in environments where a given cipher is not supported.

Modifications
-------------
When the required cipher is not supported, display a warning message saying that HTTP/2 is not being enabled and propose a solution.

Result
------
A warning message is displayed when HTTP/2 is not enabled because it's running in an unsupported environment.