SWARM-911: insufficient rules for fraction autodetection
Motivation
----------
The rules for fraction autodetection are insufficient for some
Java EE components. Most importantly, the rule for the `batch-jberet`
fraction can't possibly work, because there are no classes in the
`javax.batch` package, they are all in the subpackages.

Modifications
-------------
Modified the rules for:

- `batch-jberet`
- `jaxrs` and `jaxrs-*`
- `jpa`
- `mail`
- `transactions`
- `undertow`
- `webservices`

Result
------
More precise fraction autodetection. Autodetecting the `batch-jberet`
fraction is actually possible.