SWARM-863 - Again, fix shutdown problems on Windows
Motivation
----------

It's nice being able to control-C to stop a process on Windows.
Swarm isn't able to control-C to a stop a process on Windows.
This induces sadness.

Modifications
-------------

Instead of using Runtime#addShutdownHook() to arbitrarily shut
down our TempFileProvider at some random point after control-C
is used, explicitly use a CDI @Disposes annotation to be able
to sequence the closing of the TempFileProvider.  This is required
because SOME THINGS STILL NEED THE TEMP FILES to appropriate shutdown,
and if we delete them all willy-nilly in some arbitrary thread-scheduled
order, bad things occur.

Result
------

Control-C works on Windows more better.