SWARM-916: default security subsystem configuration different from WildFly (#258)
Motivation
----------
The default configuration of the `security` subsystem as created
by Swarm is different from WildFly's default. Most of WildFly's
default configuration also applies to Swarm, perhaps with
the exception of the `Remoting` login module.

Modifications
-------------
Modified `SecurityFraction.applyDefaults` to bring the `other`
security domain in line with WildFly default, and also to add
the `jaspitest` security domain.

Modified `EJBSecurityCustomizer` to bring the `jboss-ejb-policy`
security domain in line with WildFly default.

Added `WebSecurityCustomizer` to add the `jboss-web-policy`
security domain if missing.

Result
------
Default configuration is closer to WildFly.