SWARM-817: wildfly-swarm:run does not work for dependencies added with DependenciesContainer.addAllDependencies() (#243)
Motivation
----------

Dependencies that have been overlapping between fraction and applications have been filtered from WEB-INF/lib. This caused the application to fail.

Modifications
-------------

Introduce the a DependencyTree (buckets) that allow to reason about the origin (or owner)
of a transient dependency. If the origin is a direct project dependency, it's transient deps will remain in WEB-INF/lib

Result
------

Application can bring in dependencies on libraries as they need them and if these overlap with fractions dependencies,
they'll remain in WEB-INF/lib