SWARM-877: IllegalStateException in Topology shutdown
Motivation
----------
Shutting down a microservice that's using topology-webapp causes an IllegalStateException in the logs.

There's not a problem with that, per se, but it doesn't look very nice!

Modifications
-------------
Modified TopologyProxyService to use getOptionalValue() for Service, so as not to throw the exception if it's null, and then remove itself from further updates and return if it's null.

Result
------
No longer an exception on shutdown