SWARM-957 - Empty maven.repo.local fails ARQ tests (#271)
Motivation
----------

When CI'ing, sometimes we have an empty repository.  It should
work.

Modifications
-------------

Ensure that the JBoss Nexus Repository is part of the remote
repositories for when dependencies aren't bundled within
the uberjar and when the uberjar is executed by Arquillian.

Result
------

Using an empty maven.repo.local should work.