SWARM-918 Arquillian should be aware of the empirical context-root.
Motivation
----------

If a test deployment contains some non-/ context-root set through
jboss-web.xml within the deployment, Arquillian should be aware
of it when injecting an @ArquillianResource of URL/URI.

Modifications
-------------

Make our UberJarSimpleContainer inspect the deployment to find
a jboss-web.xml if present.  If it is present, further inspect
its contents for any <context-root> element and communicate that
through the URL Provider.

Result
------

The correct context-root is injected into Arquillian tests
if a non-standard context is used.