SWARM-940: NoClassDefFoundError for ServiceClient
Motivation
----------
ServiceClient on the constructed interface was failing to be class loaded within the deployment.

Modifications
-------------
Modified cdi-jaxrsapi to have the API classes, ServiceClient and @Service, outside in a separate API jar which the fraction then has as a dependency and defines a module.xml for.

This brings it into line with how the Netflix libraries are brought in.

Result
------
Fixed CNFE for ServiceClient