SWARM-971: Code Style verification (#284)
Motivation
----------
Currently there is no code style verification, resulting in code in the project and new PRs not necessarily conforming to the code style guidelines we've created.

Modifications
-------------
Added a checkstyle:check step to the build. Currently it doesn't fail the build as there are many components that need updating. This PR adds fixes for validation to /core pieces.

Result
------
No impact on behavior.