SWARM-919 - Not all configurables appear in the log output.
Motivation
----------

Some configuration items didn't make it to the log display of
configuration items.  This is bad.

Modifications
-------------

Introduced 2-phase scanning for configurables:

1) Scan when a configurable-holding item (such as a Fraction)
   is first produced, so that user-based properties can be
   applied as defaults.  All of these must be applied before
   any fraction customizers run because some of them look for
   the presence or absence of things to make customization
   decisions.

2) Post-customization, re-scan, to pick up anything added or
   altered by the customizers.

Phase (2) is the new addition.

Additionally, while in the code, cleaned up the config-api usage
as it relates to the SecurityFraction and the EJBSecurityCustomizer.
Prefer closures over constructors.

Since ConfigurableManager tracks objects it has scanned, also
modify it's own close() to clean up that list used to track the
objects.

Result
------

More complete display of configurable items in the log output.