SWARM-990: Sub Module classes not in right location
Motivation
----------
RuntimeDeployer assumes all paths that are dependencies are JAR files as it calls `depContainer.addAsLibrary()` for them.

However, sometimes the dependency path points to the /classes directory of another module. This is especially true when running from the IDE and it has auto added sub modules as a project dependency and not an artifact one.

Modifications
-------------
Get smarter about whether we're dealing with an actual library or a path to compiled classes, and handle them appropriately.

Result
------
Fixes problem of /classes not being added correctly for a sub module when running in the IDE.