SWARM-958 - Improve hierarchy of configuration items. (#304)
Motivation
----------

We lose the hiearchy of swarm.logging.com.foo.bar.level=TRACE
when we flatten everything down to properties, even when it was
originally represented in the YAML.

This makes it impossible to adjust logging levels through the YAML
or properties.

Modifications
-------------

Radically adjust how we handle all configuration inputs (Yaml and Properties)
to build a tree (of ConfigNodes) where each node can have a value and/or
children.

Provide a resolution strategy that can load properties or several YAMLs
and provide override capabilities.

BREAKING CHANGE: project-stages.yml is deprecated and removed from the CLI,
but if found it is still supported.  For now.

Property representation of configuration item names that include dots
(such as logger category names) are now bracketed with square brackets:

e.g.

  -Dswarm.logging.loggers.[com.foo.bar].level=TRACE

Regardless of the source of the configuration value, appropriate properties
will be set for each item, in the event someone wants to not @Inject or
otherwise deal with our fanciness, and only wish to use System.getProperty(...).

Additionally, by using -PconfigName, possibly many times, you can activate,
in-order, project-${configName}.yml as sources of inputs.  This is the
replacement to the monolithic project-stages.yml file.

These new YAML files may be found relative to $PWD or from within your
application's classpath. (not in META-INF, just at the root).

Result
------

You can now configure logging via yaml. Basically.