SWARM-1020: Fix equal position checking of Option#parse
Motivation
----------
org.wildfly.swarm.cli.Option#parse returns a specified value of a long arg with equal
like --foo=bar but it returns the value included '='(e.g. =bar) because the '=' postion
check of the method is wrong.

Modifications
-------------
The check of the '=' position is fixed apploriately.

Result
------
Option#parse returns an expected value.