SWARM-1023: Arquillian auto added to all BOMs
Motivation
----------
Arquillian is critical to be in all BOMs, so we've updated the fraction-plugin to auto add Arquillian and not include Arquillian as an auto discovered fraction when BOM building.

Also modify the build order to only build arquillian after all fractions and before testsuites.

Modifications
-------------
Reorder modules in pom.xml and upgrade to fraction-plugin 41.

Result
------
No change to product behavior.