SWARM-891: Fraction autodetection with Arquillian tests
Motivation
----------
Currently there is no way for Arquillian tests to autodetect fractions when none are specified as Maven dependencies.

Modifications
-------------
Add fraction-list to arquillian and set autodetection to `when_missing` on BuildTool.

Modify Shrinkwrap resolving to add dependencies directly as opposed to passing a list of gavs into resolve() to ensure that the scope of each dependency is set to its actual value, as Shrinkwrap defaults to compile.

Modify the ordering of explicit deps from `DependencyTree` so that compile and provided dependencies come first. This ensures that when iterating to retrieve transitives for each of them, we're retrieving transitive compile dependencies first. Otherwise a test scoped dependency could take precedence over a compile one for the same gav.

Result
------
Arquillian tests are now able to utilize autodetection of fractions.