SWARM-1005 - Delineate between project and product bits.
Motivation
----------

When/if WildFly Swarm becomes productized, there will probably
be bits that are excluded.

Modifications
-------------

Rough in profiles so that some bits can be excluded, using

  mvn install -Dswarm.product.build

Candidate bits are purely arbitrary to demonstrate that this
works, when/if it becomes necessary.

Result
------

No change unless you include -Dswarm.product.build on your
CLI, at which point, fewer things get built and included
in the fraction-lists and BOMs.