SWARM-1004 - Too many warnings when CDI is involved.
Motivation
----------

Noisy warnings in the log are noisy.

Modifications
-------------

The warning about InjectStageConfigView from [ServiceLoader]
category is actually coming from our internal Weld/CDI of
the container, not the CDI relevant to deployments.  It occurs
because when we crank up the internal Weld, it still checks
META-INF/services/... for any CDI Portable Extensions it can
find.  And it trips across the service-loader entry that is
indended only to be imported into a deployment, since it
references our extension that lives within a .deployment.*
package.

To prevent this, the TCCL is wrapped during our Weld initialization
to simply refuse to acknowledge any META-INF/services/... entries
related to CDI Portable Extensions. We explicitly set them up, so
no discovery is required.

Result
------

Less noise in the log.