SWARM-1078: Explicitly lookup maven mirror authentication
Motivation
----------
The package goal of the maven plugin fails in case a mirror is specified in settings.xml that requires authentication.

Modifications
-------------
Instead of assuming the mirror uses the same credentials as the mirrored repo, lookup and use the (optional) mirror authentication from maven session.

Result
------
Mirror is accessed with the correct credentials from settings.xml (if any).