SWARM-542: Can't run main() from IDE when using only Gradle, not Maven
Motivation
----------
When developing with gradle, it should be possible to work without the local maven repository.

Modifications
-------------
Detect when the IDE is using gradle. If gradle is used, respect the gradle cache format and use the gradle cache.

Result
------
The repository mavenLocal() isn't any more needed in the repository config.