SWARM-1063 - Do not try to process projects with pom packaging. (#348)
Motivation
----------

Some folks like to put the wildfly-swarm-plugin configuration
in a parent pom. These parent poms, though, have no artifacts
to be packaged. And thus this fails.

Modifications
-------------

If the current project has pom packaging, quietly (with a log message)
not process it.

Result
------

Should be possible to configure the plugin in a parent pom without
breaking.