SWARM-542: Can't run main() from IDE when using only Gradle, not Maven
Motivation
----------
Dependencies are resolved from the local gradle cache. But when they are missing, they are not downloaded.  Gradle doesn't download all dependencies witch are required to run wildfly swarm. So this resolver should download it by him self.

Modifications
-------------
Extended the GradleResolver class to download the missing dependencies and store them in the gradle cache. Like the DefaultMavenResolver.

Result
------
Missing dependencies are downloaded and stored in the gradle cache.