SWARM-1098: Incremental Build Support for Wildfly Swarm Gradle Plugin
Motivation
----------
The task 'wildfly-swarm-package' should only be executed, if the UberJar is not created yet,
the depending archive (jar,war) has changed or the 'swarm' extension properties for executing the task have changed.
Otherwise the task should be marked as UP-TO-DATE.

Modifications
-------------
Defining the inputs and outputs of the 'wildfly-swarm-package', so Gradle will automatically make an update-to-date check.

Result
------
The plugin will have the same behaviour on using. But running the tasks twice the second time should the task should be marked as UP-TO-DATE.