SWARM-1104 - More eagerly setup project-config.
Motivation
----------

Since people want to be about to use the ConfigView from main(),
we should intiailize it progressively so that it's useful before
swarm.start() is called.

Modifications
-------------

Initialize the config-view pieces within the Swarm() ctor and
progressively as withConfig() and withProfile() are called.

Result
------

Users may use ConfigView in main() immediately after constructing
the Swarm, or after calling withConfig() or withProfile().