SWARM-901 - Non-main() method for setting transactions ports.
Motivation
----------

Having to write a main() is bad.  It would be nice to be able to
use YAML such as

    swarm:
      transactions:
        port: 3423
        status-port: 8342

In fact, you can.

Modifications
-------------

Wrote a test to prove that ports can be set via project-defaults.yml.

Additionally added support for @DeploymentModule on @DefaultDeployment
arquillian tests to make writing such tests easier.

Result
------

200 proof port setting.