SWARM-1133: fraction autodetection broken for messaging
Motivation
----------
Ever since the fix for SWARM-974, fraction autodetection
for messaging is broken. This is because
the `MessagingPackageDetector` forgets to declare
the correct package (`javax.jms`).

Modifications
-------------
Changed the `MessagingPackageDetector` to declare
the `javax.jms` package.

Also a tiny modification to Swarmtool to avoid
printing wrong file name.

Result
------
Autodetection of the messaging fraction works again.