SWARM-1051 - Ensure jboss-deployment-structure.xml is respected.
Motivation
----------

Advanced users may wish to add dependencies to their applications
using traditional jboss-deployment-structure.xml. It should work.

Modifications
-------------

While in general jboss-deployment-structure.xml is respected, we did
have an issue where round-tripping a user-provided dependency with
the implicit slot of `main` was ignored and produced a broken
jboss-deployment-structure with a slot="" which resulted in a boot-time
issue.  This is *not* the issue reflected by the original bug report
(SWARM-1051).

Result
------

Better support for jboss-deployment-structure.xml.